#66 solution likely correct but slow even for D <= 100

+++ src/main/scala/mike/sokoryansky/MathUtils/Diophantine.scala
@@ -7,7 +7,7 @@ case class Diophantine(a: Long, xPow: Int, b: Long, yPow: Int, c: Long) {
   def y(x: Long): Option[Long] = {
     val cMinusX = c - a * Math.pow(x, xPow).toLong
     if (cMinusX % b != 0) None
-    else if (Integer.isPow(cMinusX / b, yPow)) Some(Math.pow(cMinusX / b, 1 / yPow).toLong)
+    else if (Integer.isPow(cMinusX / b, yPow)) Some(Math.pow(cMinusX / b, 1.toDouble / yPow).toLong)
     else None
   }
   def yFind(ys: Stream[Long]): Option[Long] = ys.find(y => x(y).nonEmpty)
@@ -16,7 +16,7 @@ case class Diophantine(a: Long, xPow: Int, b: Long, yPow: Int, c: Long) {
   def x(y: Long): Option[Long] = {
     val cMinusY = c - b * Math.pow(y, yPow).toLong
     if (cMinusY % a != 0) None
-    else if (Integer.isPow(cMinusY / a, xPow)) Some(Math.pow(cMinusY / a, 1 / xPow).toLong)
+    else if (Integer.isPow(cMinusY / a, xPow)) Some(Math.pow(cMinusY / a, 1.toDouble / xPow).toLong)
     else None
   }
   def xFind(xs: Stream[Long]): Option[Long] = xs.find(x => y(x).nonEmpty)
